requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Financial Libraries
yfinance>=0.2.0
ccxt>=4.0.0
quantlib>=1.30

# Risk Management & Analytics
arch>=6.0.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Telegram Bot
python-telegram-bot>=20.0
requests>=2.31.0

# Data Processing & Storage
sqlalchemy>=2.0.0
redis>=4.5.0
pymongo>=4.4.0

# Async & Concurrency
asyncio-throttle>=1.0.0
aiohttp>=3.8.0

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
mock>=5.1.0

# Data Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# ML/Forecasting (Bonus Features)
tensorflow>=2.13.0
torch>=2.0.0
prophet>=1.1.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3
websocket-client>=1.6.0
schedule>=1.2.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
